moved import-workspace to devpod pro command
