+++ cmd/pro/import_workspace.go
@@ -1,4 +1,4 @@
-package provider
+package pro
 
 import (
 	"context"
